+++ lib/langchain/tool/base.rb
@@ -29,8 +29,9 @@ module Langchain::Tool
   #
   # 3. Pass the tools when Agent is instantiated.
   #
-  #     agent = Langchain::Agent::ReActAgent.new(
-  #       llm: Langchain::LLM::OpenAI.new(api_key: "YOUR_API_KEY"), # or other like Cohere, Hugging Face, Google Palm or Replicate
+  #     agent = Langchain::Assistant.new(
+  #       llm: Langchain::LLM::OpenAI.new(api_key: "YOUR_API_KEY"), # or other LLM that supports function calling (coming soon)
+  #       thread: Langchain::Thread.new,
   #       tools: [
   #         Langchain::Tool::GoogleSearch.new(api_key: "YOUR_API_KEY"),
   #         Langchain::Tool::Calculator.new,
@@ -42,9 +43,10 @@ module Langchain::Tool
   #
   # 1. Create a new file in lib/langchain/tool/your_tool_name.rb
   # 2. Create a class in the file that inherits from {Langchain::Tool::Base}
-  # 3. Add `NAME=` and `DESCRIPTION=` constants in your Tool class
-  # 4. Implement `execute(input:)` method in your tool class
-  # 5. Add your tool to the {file:README.md}
+  # 3. Add `NAME=` and `ANNOTATIONS_PATH=` constants in your Tool class
+  # 4. Implement various methods in your tool class
+  # 5. Create a sidecar .json file in the same directory as your tool file annotating the methods in the Open API format
+  # 6. Add your tool to the {file:README.md}
   class Base
     include Langchain::DependencyHelper
 
@@ -61,44 +63,11 @@ module Langchain::Tool
       }
     end
 
-    # Returns the DESCRIPTION constant of the tool
-    #
-    # @return [String] tool description
-    def description
-      self.class.const_get(:DESCRIPTION)
-    end
-
-    # Sets the DESCRIPTION constant of the tool
-    #
-    # @param value [String] tool description
-    def self.description(value)
-      const_set(:DESCRIPTION, value.tr("\n", " ").strip)
-    end
-
-    # Instantiates and executes the tool and returns the answer
-    #
-    # @param input [String] input to the tool
-    # @return [String] answer
-    def self.execute(input:)
-      warn "DEPRECATED: `#{self}.execute` is deprecated, and will be removed in the next major version."
-
-      new.execute(input: input)
-    end
-
     # Returns the tool as a list of OpenAI formatted functions
     #
     # @return [Hash] tool as an OpenAI tool
     def to_openai_tools
       method_annotations
-    end
-
-    # Executes the tool and returns the answer
-    #
-    # @param input [String] input to the tool
-    # @return [String] answer
-    # @raise NotImplementedError when not implemented
-    def execute(input:)
-      raise NotImplementedError, "Your tool must implement the `#execute(input:)` method that returns a string"
     end
 
     # Return tool's method annotations as JSON
@@ -110,17 +79,6 @@ module Langchain::Tool
           self.class.const_get(:ANNOTATIONS_PATH)
         )
       )
-    end
-
-    # Validates the list of tools or raises an error
-    #
-    # @param tools [Array<Langchain::Tool>] list of tools to be used
-    # @raise [ArgumentError] If any of the tools are not supported
-    def self.validate_tools!(tools:)
-      # Check if the tool count is equal to unique tool count
-      if tools.count != tools.map(&:name).uniq.count
-        raise ArgumentError, "Either tools are not unique or are conflicting with each other"
-      end
     end
   end
 end